move MappingResolver to avoid ClassCastException on wildfly
