+++ instrumentation-api/src/main/java/io/opentelemetry/instrumentation/api/servlet/MappingResolver.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.servlet.naming;
+package io.opentelemetry.instrumentation.api.servlet;
 
 import java.util.ArrayList;
 import java.util.Collection;
