+++ instrumentation/servlet/servlet-common/library/src/main/java/io/opentelemetry/instrumentation/servlet/naming/ServletMappingResolverFactory.java
@@ -5,6 +5,7 @@
 
 package io.opentelemetry.instrumentation.servlet.naming;
 
+import io.opentelemetry.instrumentation.api.servlet.MappingResolver;
 import java.util.Collection;
 
 public abstract class ServletMappingResolverFactory {
